cleanup: job: rename RestoreRepository to appropriate name
In general, `Repository` should be used to provide simple
CRUD interfaces. However, `RestoreRepository` isn't fit
for purpose because it includes functions for complicated
I/O process. Let's rename this to `RestoreVolume`.

NOTE:
`RBDRepository` also has the same pain point. It has
`ApplyDiff` method for applying diff data to the raw.img.
I'll keep this problem as is to keep this PR as simple
as possible and will revisit this problem later.

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>