finbackup: watch job completion correctly
We shouldn't use ownerReference and `Owns()` to watch
the completion of the backup job because cross-namespace
references might happen. We should use `Watches()` instead.

We should define a map function and a predicate.

- A map function to enqueue FinBackup corresponding
  to the backup job.
- A predicate to enqueue FinBackup only if the backup
  job has just completed.

For more information about map function and predicate,
please refer to the following document.

map function:
https://book.kubebuilder.io/reference/watching-resources/secondary-resources-not-owned

predicate:
https://book.kubebuilder.io/reference/watching-resources/predicates-with-watch

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>