infra: db: reject multiple rows in backup_metadata
The current implementation of infra/db still works if the
backup_metadata table contains more than one row. However, this is an
erroneous situation, and we should detect and handle it rather than
allow it to go unnoticed.

This commit addresses the issue by making GetBackupMetadata() return an
error if the backup_metadata table has more than one row.