make both FinBackup and FinRestore possible on arbitrary nses
Both FinBackup and FinRestore can belong to arbitrary namespaces.
Howerver, these resources are forced to exist on the same
namespace as rook/ceph cluster. Let's fix it.

As a side effect of this change, I found that the unit tests
for controllers for both resources' controllers has always
succeeded due to the above mentioned namespace restriction.
In other words, these tests were virtually noop.

Let's remove these tests because the purpose of these tests are
just to verify running reconcilers without error and this scenarios
are covered by the existing e2e tests.

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>