e2e: unify backup and restore suite
Many proceses are the same or very similar in the backup
and the restore suites. Let's unify these suites.

Originally the backup suite creates Fin resources
in the rook namespace. On the other hand, the restore suite
creates these resources in the same namespace as the backup
target PVC and creates the restore PVC in the rook namespace.
It's better to fix this inconsistency as follows:

- Create Fin resourses in the rook namespace. In other words,
  the different namespace than the backup target PVC's one.
- Create the restore PVC in the same namespace as the backup
  target PVC.

This combination is a typical for Fin.

NOTE:
Probably creating another dedicated namespace for Fin
resources is better than repurposing the rook namespace.
I'll fix this problem in a different commit.

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>