e2e: try to load both kvm modules for Intel and AMD
We have been suffered from flaky EPERM error on loading
kvm_amd kernel module. It's because GitHub runner can be
run on both Intel and AMD machines. To resolve this problem,
try to load both kvm modules for intel and AMD.

An example when a runner is running on top of Intel mahine:
https://github.com/cybozu-go/fin/actions/runs/18839463075/job/53748050527#step:3:5
> Run grep -E 'svm|vmx' /proc/cpuinfo
> flags		: <snip> vmx <snip>

On top of AMD machine:
https://github.com/cybozu-go/fin/actions/runs/18839463075/job/53748139582#step:3:5
> flags		: <snip> svm <snip>

Typically one of both are automatically loaded beforehand by OS.
Howerver, it's better to load these modules explicitly.

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>